+++ src/js/constants.js
@@ -35,6 +35,7 @@ export const EVENT_DRAG_START = 'dragstart';
 export const EVENT_FOCUSIN = 'focusin';
 export const EVENT_KEY_DOWN = 'keydown';
 export const EVENT_LOAD = 'load';
+export const EVENT_ERROR = 'error';
 export const EVENT_TOUCH_END = IS_TOUCH_DEVICE ? 'touchend touchcancel' : 'mouseup';
 export const EVENT_TOUCH_MOVE = IS_TOUCH_DEVICE ? 'touchmove' : 'mousemove';
 export const EVENT_TOUCH_START = IS_TOUCH_DEVICE ? 'touchstart' : 'mousedown';